requirements.txt
pandas>=2.3.1
sentence-transformers>=5.1.0
faiss-cpu>=1.11.0.post1
openai>=1.99.3
dotenv-python>=0.0.1